Order added

+++ src/main/java/azcompany/final_project/service/impl/CartItemServiceImpl.java
@@ -10,10 +10,12 @@ import azcompany.final_project.model.dto.response.CartItemResponse;
 import azcompany.final_project.model.entity.BookEntity;
 import azcompany.final_project.model.entity.CartEntity;
 import azcompany.final_project.model.entity.CartItemEntity;
+import azcompany.final_project.model.entity.UserEntity;
 import azcompany.final_project.repository.BookRepository;
 import azcompany.final_project.repository.CartItemRepository;
 import azcompany.final_project.repository.CartRepository;
 import azcompany.final_project.service.abstracts.CartItemService;
+import azcompany.final_project.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -26,12 +28,14 @@ public class CartItemServiceImpl implements CartItemService {
     private final CartRepository cartRepository;
     private final CartItemRepository cartItemRepository;
     private final CartItemMapper cartItemMapper;
+    private final SecurityUtil securityUtil;
 
     @Override
-    public CartItemResponse addItem(Long cartId, CartItemAddRequest cartItemAddRequest) {
+    public CartItemResponse addItem(CartItemAddRequest cartItemAddRequest) {
         log.info("CartItemService.addItem.start: {}", cartItemAddRequest);
-        CartEntity cartEntity = cartRepository.findById(cartId)
-                .orElseThrow(() -> new NotFoundException("Cart not found with id: " + cartId));
+        UserEntity user = securityUtil.getUser();
+        CartEntity cartEntity = cartRepository.findByUser(user)
+                .orElseThrow(() -> new NotFoundException("Cart not found"));
 
         BookEntity book = bookRepository.findById(cartItemAddRequest.getBookId())
                 .orElseThrow(() -> new RuntimeException("Book not found with id: " + cartItemAddRequest.getBookId()));
@@ -63,6 +67,7 @@ public class CartItemServiceImpl implements CartItemService {
         log.info("CartItemService.getItemById.start: {}", id);
         CartItemEntity entity = cartItemRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Cart not found with id: " + id));
+        securityUtil.validateAccess(entity.getCart().getUser());
         CartItemResponse cartItemResponse = cartItemMapper.toResponse(entity);
         log.info("CartItemService.getItemById.end: {}", cartItemResponse);
         return cartItemResponse;
@@ -73,6 +78,7 @@ public class CartItemServiceImpl implements CartItemService {
         log.info("CartItemService.updateItemById.start: {}", id);
         CartItemEntity entity = cartItemRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Cart not found with id: " + id));
+        securityUtil.validateAccess(entity.getCart().getUser());
         if (request.getQuantity() >= entity.getBook().getStockCount()) {
             throw new QuantityException("Quantity is greater than stock count");
         }
@@ -88,6 +94,7 @@ public class CartItemServiceImpl implements CartItemService {
         log.info("CartItemService.deleteItemById.start: {}", id);
         CartItemEntity entity = cartItemRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Cart not found with id: " + id));
+        securityUtil.validateAccess(entity.getCart().getUser());
         cartItemRepository.delete(entity);
         log.info("CartItemService.deleteItemById.end: {}", entity);
     }
